Prevent an exponential behaviour in lambda-erasure in extraction.
We factor out the same call to the recursive function [kill_dummy],
which is fine because this function cannot raise an Impossible
exception. The optimization is sound because basically for any
term c,

kill_dummy (kill_dummy_lams (kill_dummy c)) = kill_dummy_lams (kill_dummy c).

The algorithm is still quadratic or so, but the size of the proof terms
shouldn't be that big to be a problem in practice.

Fixes #17448: extraction optimizations performance issue.