Added a new projection generation for congruence
The new projection generation can now generate projections for some dependently typed fields.
For this it looks if the type can be expressed as an applicative term consisting only of terms present in the environment or that can be extracted from parameters or indices of the inductive type that is getting projected.
One issue I still have is with universes. Even though the generation isn't able to handle match terms right now it should be possible when the exact match statement is extracteable from an inductive index. When I try this however it doesn't work because of something with the universes of the terms.