Remove the notation for "only" in Ltac2.
This notation is not well-bracketed and does not play well with
developments that would use "only" as a normal identifier. There is
furthermore little point in adding it as a notation since it is basically
just a call to Control.focus. The reason we had it in Ltac1 was that
everything had to go through notations, it is more than time not to
repeat the same error in Ltac2.

Since it was added surreptiously in #20197 which is not part of a released
version of Rocq, we can just remove it bluntly.