Tweak the output of the test-suite logs.
Since the Makefile test-suite does not seem to want to go
away in the immediate future, this commit adds a small
life-improving tweak to the output of the test-suite logs.
Basically, we just add a space after the file name so
that it becomes easy to select it in interfaces that split
words on spaces. For good typographical measure, we also
add a space after the trailing dots.