Tweak the native compilation scheme of cofixpoints.
Instead of conflating the application of functional cofix values with the
forcing of the node, we implement cofixes directly as a loop in memory and
rely on the fact that they are always passed as fully η-expanded functions
by the Genlambda caller.

This is more efficient as we get more sharing and perform less applications
when forcing cofixes.