Fix returning delimiter instead of scope name for About in specific scope
interp_notation_as_global_reference_expanded returns the scope name,
gotten from the key in the scope map.

When checking a specific scope (`About "foo"%bar`) it is specified
using it delimiter, and we would produce a map restricted to that
scope but incorrectly using the delimiter instead of the scope name.

Then the delimiter would be returned and used as a scope name, leading
to errors.

Fix #20754