Correctly declare functor bodies in subtyping checks.
When checking functors, one has to add the contents of the body in the
environment before proceeding to check what is inside. Indeed, the
algorithm expects the functor definitions to be in scope.

The previous code was doing something extremely weird, trying to declare
this in a place that made no real sense. Furthermore it was overwriting
the mapping of the functor with its body, resulting in a double declaration
of the same module. Worse, this was performed at every functor parameter.

We fix this by simply pushing the subcomponents of a functor in the local
environment when reaching the base case. We do not overwrite the functor
definition since this binding is not needed to check its contents, as
the graph of dependency is acyclic.