Do not redeclare functor parameters in Declaremods.
The code is extremely quirky, we keep playing mutably with the global
environment without a lot of respect for the invariants of the module
system. This patch works around a double declaration of functor parameters,
although it is not optimal. We rebuild the module type out of the
parameters known to be in the environment. If you ask me, we should have
a more principled approach but this requires heavy reengineering of
the Declaremods implementation.