Ignore module redeclaration in extraction.
Extraction has a strong tendency to overwrite modules
by their signatures, which is breaking a lot of implicit
invariants. It is not easy to fix without reimplementing
the module extraction from scratch. All bets are off in
this case, so we introduce a new internal API in Environ
to cover this extremely dubious behaviour.