Do not keep readding modules when computing self inclusions.
We only care about the current prefix of the module being registered
as a module in the environment, because at the call point the module
is still open and thus its subcomponents are part of the environment.
Given that the code was basically copy-pasted, both the kernel and
Declaremods must be fixed the same way.