Prevent module redeclaration in module type subtyping.
The rule to check field subtyping when the field is a module type
boils down to checking that the two modules types are equivalent.
The previous code had an ad-hoc check for this, but to satisfy
that the module contents are part of the current environment, when
checking mp₁ : M₁ ≅ mp₂ : M₂, it was adding both module paths
to the environment. Unfortunately, there is no requirement that
the two pathes are distinct, and in practice some callers do
pass the same path on both sides. The net result is that one
implementation took precedence over the other.

I am not sure this can be weaponized into an unsoundness proof,
but it clearly messes with some other expected invariants. To fix
this, we remove the ad-hoc check for module equivalence and replace
it with a subtyping check in both directions. This is not really
worse algorithmically-wise, because the previous check also ended
up performing a two-way check when it reached non-functor modules.
Hence both implementations are exponential in the depth of nested
module types. The only difference is that the previous code was
a bit more efficient with module parameters, but in practice it
should not matter that much, and it is also unclear to me that
code in the wild stores functor types (!) in modules, let alone
deeply nested.