Do not overwrite modules in native compilation.
We silently change the expected preconditions of Nativelibrary.dump_library
function. Instead of registering the contents of the module being compiled
ourselves, we let the caller do this. One of the two callers did already
register the module contents, so since we cannot unregister it, we also
require the other caller to do it beforehand.