Do not overwrite modules when extracting abstract subcomponents.
This code path is only taken when extraction is targetting another language
than OCaml. It tries to expand away the contents of the module, but when
facing abstract modules such as bound parameters, it had the unfortunate
side-effect of overwriting the module binding. We fix this by expanding
it to a dummy name that is substituted away right after.