Properly update resolvers of submodules when performing strengthening.
An implicit invariant of delta-resolvers is that when nesting non-functor
modules, the resolver associated to an inner module is equal to one obtained
by taking the resolver of the encompassing module and restricting it to the
submodule modpath. Said otherwise, resolvers from inner modules are
completely redundant.

The previous code was losing the original resolver from the inner module when
performing aliasing calls like e.g. Module M := N, leading to a breakage of
this implicit invariant. This was leading to further silent issues down the
road but since the module code is not well-excercised, this flew under the
radar.