Fix substitution of functor delta-resolvers when performing strengthening.
The previous code was only substituting the inner delta resolvers and ignoring
the codomain of functors. In particular this was generating ill-formed constants
whose canonical component was pointing to a bound name that did not exist in the
global environment. We fix the code by also substituting the codomain of module
bodies. This is actually in line with what we do in the related function
strengthen_and_subst_module_body, where we were also correctly substituting the
codomain.

This is a quite old soundness bug that was also affecting the checker. It seems
to have been introduced by c5b699f8 by an oversight when refactoring tricky code.

Fixes #21051: Unsoundness with heavy module machinery.