Check for coherence of primitive constants using user names.
The previous code was relying instead on canonical names. While there
may be weird situations where aliasing a module might generate primitive
operations which is non-canonical, I believe this is actively looking for
trouble. See also #14822 for a lengthier argumentation. The new code is
stricter than the previous one, so we are erring on the safe side. If
ever somebody comes up with a crazy scheme requiring to have aliased
primitive operations we may reconsider, but for the time being I am
convinced this patch is a net enhancement.