Use user names to cache computation of inductive case tables in Genlambda.
This is very unlikely to make any difference, it should be quite rare to
mix inductive types with the same canonical name but different user names
in the same term. At any rate I am not even sure this cache is really
useful and not just some premature optimization, I never see this popping
anywhere.