Reimplement abstract in a purely functional way.
This patch replaces the infamous mutable modification of the global
safe environment in the abstract tactic by a purely functional mechanism.
This changes some of the hidden invariants and makes it a non-trivial
change likely to uncover some subtle issues in the implementation. I tried
to make it as lenient as possible as a first pass, we will try to sanitize
the downstream clients later on.

The main change is that an environment alone is not enough to make sense
of constants from terms living in the tactic layer. Local constants defined
by abstract are not part of the environment but should rather be obtained
through the evarmap. Note that this is actually somewhat innocuous API-wise,
since the EConstr API has been precisely designed to always require an
evarmap around. The only changes are indirect callers to lookup_constant,
which must now be given an evarmap.

A more subtle change is the precise relation between the global and tactic
environments. An explicit design choice is that between vernacular commands,
the global environment does contain the abstracted constants. For this
to work, we need a two-tiered side-effect tracking. The first tier is
provided by the evarmap, which only contains the local constants that have
been generated by the current tactic invocation. The second tier is
implemented in the Declare mechanism, where we track all private constants
defined by the previous tactic calls in the current proof in order to replay
them when closing it.