Use user name equality in checker rtree test.
The use of canonical equality was doubly wrong. First, this test is only
to check that the recomputation of an inductive block by the checker is the
same as the original one, so the translation process should never introduce
module aliases. Second, this precise check only occurs whenever nesting
primitive arrays, and anybody coming up with a inductive type nesting over
aliased primitive arrays and expecting it to be fine should seriously
reconsider their life choices.