Rely on user ordering for assumption sets and maps.
I believe that the previous implementation was wrong, because it quotients
out aliased names in the result of Print Assumptions and friends. This is
clearly wrong for transparency analysis, and very dubious for unsafe feature
analysis. The way we represent the assumption data is still quite wonky in
my opinion, but this patch is a minimal way to get rid of the reliance on
canonical names while retaining the previous implementation.