Use user equality in comparison of native values.
This is a slight change of semantics, since it means aliased constants will
be compiled separately. In practice it probably doesn't matter because most
aliased constants have a body that is exactly the constant they alias, so
the only difference is that we will recompile definitions that have the same
body up to aliasing. This should be very rare for native use cases though.