Cache the result of node expansion of inductive guard checking.
For big mutually recursive types, we kept recomputing the recarg
expansion again and again. This was a clear source of inefficiency
as witnessed in #21105. Since in this case the data is uniquely
given by the inductive block, we simply cache it once and for all.

Fixes #21105: x10 performance regression in guard check.