Constant effects in Reductionops are now registered per user name.
This is a half-baked feature that is only used by some obscure
libraries to hack into the upper-layer reduction machine in order
to perform side-effects such as displaying stuff to the user. In
addition to not being observable due to the lack of users, I would
argue that if you are registering effects up to aliasing you deserve
whatever happens.