Rely on user names for projection table in extraction.
The previous code was probably buggy, because the corresponding
inductive table uses user names, so there might have been a
desynchronization. That said, it is not even clear because the data
is the same for aliased inductive types, and was not even accessed
by the code. So this change should be a no-op. To make this clearer,
we seize the opportunity to replace maps with sets.