Do not rely on term ordering in Search priority ordering.
The code was actually only ever considering global references. This
patch is a slight change of semantics for two reasons. First, the
previous code was also taking into account the universe instances
of the references, which would have resulted in a dubious semantics
for universe polymorphic code. Second, we now consider aliased
references distinct. In any case, this piece of code is just a
heuristic for a printing order, so it should not matter.