Improved relevance handling in eqschemes
This changes the error in #20099 from "bad binder relevance" to

~~~
Error:
Incorrect elimination of "H" in the inductive type "eq@{foo.34}":
the return type has sort "Type@{foo.35}" while it should be SProp.
Elimination of an inductive object of sort SProp
is not allowed on a predicate in sort "Type@{foo.35}"
because strict proofs can be eliminated only to build strict proofs.
~~~