Fix ssrintro
The `tactic => intro_pattern` ssreflect notation was moved from
notation level 1 to notation levels 3 and 4 in
https://github.com/rocq-prover/rocq/commit/39131994a85de4cd2edcb27bbed8428c8301cae9
in order to minimize use of the recovery mode. This still wasn't fully
satisfying as seen in https://github.com/rocq-prover/rocq/pull/21126
This moves it back to level 1l, just below 1 but left associative
(level 1 is right associative).