Add "Ltac2 Import Type" command
This should be useful if we want to move a type up in the dependency graph.

eg if we want to move Std.reference to Init, just doing
`Ltac2 Type reference := Init.reference` in Std is not enough for
backwards compat since the constructor (Std.ConstRef etc) would not be
available from Std.

Making notations is also not enough as they won't be used in patterns.

With this new feature we can instead do
`Ltac2 Import Type Init.reference as reference.`
and `Std.ConstRef` will refer to `Init.ConstRef` etc.