Collapse NonLogical thunks in the logic monad.
We leverage the fact that a ~> unit -> b ≅ a -> b where ~> stands
for a hypothetical pure arrow type. This allows replacing all
instances of the NonLogical.t monad in the logic monad type with
basically nothing, leaving effects implicits in the rightmost
arrow type. Since all clients evaluate the thunk directly, the
new code should be equivalent to the previous one.

Actually, it may even be more correct given that we already
implicitly use the function space in the monadic bind operator
to perform side-effects.