Refine the cache check in TC search algorithm.
The previous code was computing the domain of the sort elimination graph,
which is linear in the size of this set and thus very costly for a cache.
We rely instead on physical equality of the graph itself, which seems to
empirically coincide with the former but is O(1).

This source of slowness was observed in the wild.