Make the failed-abstract-certificate warning error by default.
Since the purification of the side-effect mechanism, this warning should
never be triggered. We could make it into an anomaly, but just for the
sake of being able to work around bugs, we just make it error by default
from now on, and will eventually turn it into an assertion.