Do not wrap side-effect certificates in ephemerons.
The reason we used to do so was likely because at some point safe
environments contained futures and thus arbitrary closures, that could not
be marshalled safely. Nowadays this datatype is first-order, so there
should not be any problem on this front. Note that despite marshalling
losing pointer equality, the current implementation of par: is done in
a way to ignore abstract wrappers and replay the proof locally without
marshalling the certificate, thus preserving its validity.

The only potential issue comes from the huge size of the marshalled
environment in async mode, but in theory we already marshal it anyway
in the summary, so most of the data should be shared.