Make side-effects globals to the proof block that generated them in Declare.
This cleans up a longstanding issue with effect deduplication that made
no sense whatsoever. The new model has better static invariants, as it
separates the first phase where effects are common to the block of
constants being defined, and the second phase where we decide whether
to inline them on a per-constant regard based on opacity.

This code should not change the semantics, it is just a clean-up for now.