Correctly use the goal-local evarmap to check toplevel TC unifiability.
The code for toplevel unifiability was not performing the check in the
same evarmap as the one performed during resolution. The former was using
the evarmap from the point where the goals depended upon were fetched,
while the latter used the evarmap from the current goal being solved.

According to the oracle Sozeau, we should look at unifiability in the
current goal, so we fix the toplevel check.