Reintroduce Univ.ContextSet.
It makes sense to have a type that only contains universe constraints
and not elimination constraints for monomorphic code. As a matter of
fact, PConstraints.Context.Set.t is extremely suspect as it allows
to add sort constraints but no sort variables, an asymmetry that should
have raised a red alarm.