Some additional safeguards around dropping of elimination constraints.
Weirdly enough, just asserting that the set of elimination constraint
is empty in monomorphic definitions is enough to pass the test-suite.
If you ask me this is fishy, as it means we never really generate
any constraint in this mode. At best we should generate constraints
on local sort variables and then collapse them to trivially true
ground constraints. This probably means somebody will have to fix
things up when we really start using sort polymorphism.