Perform kernel typing after universe abstraction for polymorphic code.
This is safer and will make it possible to simplify the code that pushes
bound levels and sorts in the environment. We also share more hashconsing
for polymorphic entries this way.

This also probably fixes a bug with the relevance computation for definitions.
The environment passed to the function was the one with the concrete levels,
but the term argument was expected to live in the abstracted context.