Autorewrite uses a more principled universe context type.
The UState.merge_sort_context_set function is stricter than
UState.merge_context_set, as the latter does not check anything about
the elimination constraints but simply pushes them in the local
state. We have to hack our way around to emulate the old behaviour,
but at least the issue is made explicit.