Fix #21434: coqchk gives fatal error on module subtyping.
The checker was incorrectly ignoring the δ-resolver associated to the
argument of a functor application when recomputing the signature of
an algebraic expression, leading to ill-formed kernel pairs. We fix
it by doing basically the same thing as the kernel, up to the fact
that we do not try to perform inlining. It should not matter as typing
is stable by conversion.