Centralize the Nativenorm global state in a reference-passed record.
This was a very bad implementation that is, for some reason, common in
the OCaml compiler world. Rather than haphazardly mutating some global
references, with pass instead a locally mutable state along and ask
toplevel callers to provide one.

We move the code piecewise to reduce the amount of potential breakage.
This commit only moves the global table to the local state.