Do not allow declaring global sorts inside sections.
Typical "It hurts when I bang my head against the wall." moment. This
is a big vector attack that makes the code much more complex for no
good reason. Indeed, contrarily to levels that are added by virtually
every command, adding a sort should be an extremely rare endeavour.
If the user really wants it, they can just hoist the declaration out
of the section.