Use sets rather than lists in Evarnames internals.
This is algorithmically much better. The only observable difference is the
name associated to an evar by Evd.evar_ident in case of conflict. Looking at
the callers, it is not very clear what is really expected in practice, and
the previous code was not given a lot of thought about this situation. We
should probably let the caller decide themselves, but for the sake of
backwards compatibility of the API, in this commit we silently change this
marginal behaviour and will fix the API at a later point.