Move sort cumulativity constraint types out of the kernel.
The only raison d'être of those types in addition to elimination constraints
is to handle the Prop ⊆ Type rule in unification. They have no use in the
kernel, so we move them to the most reasonable place in the engine where
they start to make sense, i.e. UnivProblem.